Created separated files for structs and classes
