+++ Sources/ta4swift/Strategy/Rule.swift
@@ -2,47 +2,10 @@
 //  File.swift
 //  
 //
+//  Created by Simon-Justus Wimmer on 23.01.22.
 //
 
 import Foundation
-
-
-public protocol Strategy {
-    
-    var entryRule: Rule { get }
-    var exitRule: Rule { get }
-    var unstablePeriod: Int { get set}
-    
-    func isUnstableAt(index: Int) -> Bool
-    func shouldEnter(index: Int, record: TradingRecord) -> Bool
-    func shouldExit(index: Int, record: TradingRecord) -> Bool
-}
-
-public struct BaseStrategy: Strategy {
-    
-    public let entryRule: Rule
-    public let exitRule: Rule
-    
-    public var unstablePeriod = 0
-    
-    public init(entryRule: Rule, exitRule: Rule) {
-        self.entryRule = entryRule
-        self.exitRule = exitRule
-    }
-    
-    public func isUnstableAt(index: Int) -> Bool {
-        return unstablePeriod > index
-    }
-    
-    public func shouldEnter(index: Int, record: TradingRecord) -> Bool {
-        return isUnstableAt(index: index) && entryRule.isSatisfied(for: index, record: record)
-    }
-    
-    public func shouldExit(index: Int, record: TradingRecord) -> Bool {
-        return isUnstableAt(index: index) && exitRule.isSatisfied(for: index, record: record)
-    }
-    
-}
 
 public protocol Rule {
     
@@ -143,24 +106,6 @@ public struct XOrRule: Rule {
     
     public func isSatisfied(for index: Int, record: TradingRecord) -> Bool{
         return rule1.isSatisfied(for: index, record: record) ^ rule2.isSatisfied(for: index, record: record)
-    }
-    
-}
-
-public struct BooleanRule: Rule {
-    
-    public let satisfied: Bool
-    
-    public init(_ satisfied: Bool) {
-        self.satisfied = satisfied
-    }
-    
-    public func isSatisfied(for index: Int) -> Bool {
-        return satisfied
-    }
-    
-    public func isSatisfied(for index: Int, record: TradingRecord) -> Bool {
-        return satisfied
     }
     
 }
