+++ Sources/ta4swift/TradingRecord/Position.swift
@@ -1,0 +1,13 @@
+//
+//  File.swift
+//  
+//
+//  Created by Simon-Justus Wimmer on 24.01.22.
+//
+
+import Foundation
+
+public struct Position {
+    var entry: Trade
+    var exit: Trade
+}
