+++ Sources/ta4swift/TradingRecord/Runner.swift
@@ -1,0 +1,17 @@
+//
+//  File.swift
+//  
+//
+//  Created by Simon-Justus Wimmer on 24.01.22.
+//
+
+import Foundation
+
+public struct Runner {
+    
+    func run(strategy: Strategy, type: TradeType) -> TradingRecord {
+        let record = BaseTradingRecord();
+        
+        return record
+    }
+}
