+++ Sources/ta4swift/TradingRecord/Trade.swift
@@ -1,0 +1,14 @@
+//
+//  File.swift
+//  
+//
+//  Created by Simon-Justus Wimmer on 24.01.22.
+//
+
+import Foundation
+
+
+public struct Trade {
+    let type: TradeType
+    let index: Int
+}
