+++ Sources/ta4swift/TradingRecord/TradingRecord.swift
@@ -1,0 +1,24 @@
+//
+//  TradingRecord.swift
+//  
+//
+//
+
+import Foundation
+
+
+public enum TradeType {
+    case BUY
+    case SELL
+}
+
+public protocol TradingRecord {
+    
+}
+
+public struct BaseTradingRecord: TradingRecord {
+    
+    public init() {
+        
+    }
+}
